requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
Babel==2.9.1
Baker==1.3
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
boto3==1.26.32
botocore==1.29.32
Brlapi==0.8.3
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.1.4
django-allauth==0.50.0
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.1
django-cors==0.1
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-nine==0.2.7
django-phonenumber-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.1
drf-yasg==1.21.4
duplicity==0.8.21
elastic-transport==8.4.0
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
exceptiongroup==1.0.4
Faker==15.3.4
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.3.3
fs==2.4.16
future==0.18.2
google-api-core==2.8.2
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
gpg===1.16.0-unknown
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.4
humanize==4.4.0
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcomps==0.1.15
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
magic-filter==1.0.9
Mako==1.1.3
MarkupSafe==2.1.1
model-bakery==1.9.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.3
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==22.0
paramiko==2.9.3
pexpect==4.8.0
phonenumbers==8.12.47
Pillow==9.3.0
pipenv==2022.12.19
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.21.1
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.0
pytest-django==4.5.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
pyxdg==0.27
PyYAML==6.0
redis==4.3.4
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
rpm==4.17.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.0.1
sqlparse==0.4.3
stripe==4.1.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
tornado==6.2
twilio==7.9.3
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2022.6
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.13
usb-creator==0.3.7
users==1.0.dev0
vine==5.0.0
virtualenv==20.13.0+ds
virtualenv-clone==0.5.7
wadllib==1.3.6
wcwidth==0.2.5
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
